workspace-colors: give the allocation rules a constructable owner
WorkspaceAutoTabColorAssignment was a caseless enum holding nothing but
static helpers, which is the ambient global surface no-ambient-global-state
rules out for new production code. The type is gone and each piece moved to
whatever already owned that decision.

nextColorHex had no production caller left once reconcile allocated from
one allocator, so it is deleted rather than moved. The tests that used it
build an allocator through a local helper, which leaves exactly one
allocation entry point shipping.

railColorHex becomes an instance method on WorkspaceIndicatorStyle, next to
automaticallyAssignsWorkspaceColors. The style is what decides whether a
rail is drawn, so the enablement rule now reads off the value that owns it.

normalized becomes String.workspaceColorKey, so comparing two hexes is a
property of the hex rather than a call into a helper bag.
