+++ Packages/macOS/CmuxSettings/Sources/CmuxSettings/Values/WorkspaceIndicatorStyle.swift
@@ -14,6 +14,25 @@ public enum WorkspaceIndicatorStyle: String, CaseIterable, Sendable, SettingCoda
         self == .leftRailAuto
     }
 
+    /// Resolves the rail color for one workspace, applying the full enablement
+    /// rule for this style.
+    ///
+    /// Returns `nil` — leaving the rail hidden exactly as before the automatic
+    /// style existed — whenever this style does not assign colors, the
+    /// workspace already carries a manual color, or no color has been assigned.
+    ///
+    /// An empty `customColorHex` counts as no manual color, matching the
+    /// allocator, which also treats it as uncolored and hands the workspace an
+    /// automatic color. Disagreeing here would assign a color and then refuse
+    /// to draw it.
+    public func railColorHex(customColorHex: String?, assignedColorHex: String?) -> String? {
+        guard automaticallyAssignsWorkspaceColors,
+              customColorHex?.isEmpty != false else {
+            return nil
+        }
+        return assignedColorHex
+    }
+
     /// Maps raw strings written by earlier iterations of the indicator
     /// setting onto the closest modern case, exactly as the legacy
     /// `SidebarActiveTabIndicatorSettings.resolvedStyle` did. Unknown
